Struct google_groupssettings1::Groups

pub struct Groups {
    pub who_can_ban_users: Option<String>,
    pub who_can_assist_content: Option<String>,
    pub allow_external_members: Option<String>,
    pub who_can_enter_free_form_tags: Option<String>,
    pub who_can_approve_messages: Option<String>,
    pub who_can_mark_duplicate: Option<String>,
    pub who_can_join: Option<String>,
    pub who_can_modify_tags_and_categories: Option<String>,
    pub who_can_mark_no_response_needed: Option<String>,
    pub who_can_unmark_favorite_reply_on_any_topic: Option<String>,
    pub who_can_moderate_content: Option<String>,
    pub primary_language: Option<String>,
    pub who_can_mark_favorite_reply_on_own_topic: Option<String>,
    pub who_can_view_membership: Option<String>,
    pub favorite_replies_on_top: Option<String>,
    pub who_can_mark_favorite_reply_on_any_topic: Option<String>,
    pub include_custom_footer: Option<String>,
    pub who_can_move_topics_out: Option<String>,
    pub default_message_deny_notification_text: Option<String>,
    pub include_in_global_address_list: Option<String>,
    pub archive_only: Option<String>,
    pub who_can_delete_topics: Option<String>,
    pub who_can_delete_any_post: Option<String>,
    pub is_archived: Option<String>,
    pub members_can_post_as_the_group: Option<String>,
    pub who_can_make_topics_sticky: Option<String>,
    pub custom_roles_enabled_for_settings_to_be_merged: Option<String>,
    pub email: Option<String>,
    pub who_can_discover_group: Option<String>,
    pub who_can_modify_members: Option<String>,
    pub message_moderation_level: Option<String>,
    pub description: Option<String>,
    pub who_can_unassign_topic: Option<String>,
    pub reply_to: Option<String>,
    pub custom_reply_to: Option<String>,
    pub send_message_deny_notification: Option<String>,
    pub enable_collaborative_inbox: Option<String>,
    pub who_can_contact_owner: Option<String>,
    pub message_display_font: Option<String>,
    pub who_can_leave_group: Option<String>,
    pub who_can_add: Option<String>,
    pub who_can_post_message: Option<String>,
    pub who_can_move_topics_in: Option<String>,
    pub who_can_take_topics: Option<String>,
    pub name: Option<String>,
    pub kind: Option<String>,
    pub max_message_bytes: Option<i32>,
    pub who_can_invite: Option<String>,
    pub who_can_approve_members: Option<String>,
    pub spam_moderation_level: Option<String>,
    pub allow_web_posting: Option<String>,
    pub who_can_moderate_members: Option<String>,
    pub who_can_add_references: Option<String>,
    pub who_can_view_group: Option<String>,
    pub show_in_group_directory: Option<String>,
    pub who_can_post_announcements: Option<String>,
    pub who_can_lock_topics: Option<String>,
    pub who_can_assign_topics: Option<String>,
    pub custom_footer_text: Option<String>,
    pub allow_google_communication: Option<String>,
    pub who_can_hide_abuse: Option<String>,
}

JSON template for Group resource

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

update groups (request|response)

patch groups (request|response)

get groups (response)

Fields
who_can_ban_users: Option<String>

Permission to ban users. Possible values are: NONE OWNERS_ONLY OWNERS_AND_MANAGERS ALL_MEMBERS

who_can_assist_content: Option<String>

Permission for content assistants. Possible values are: Possible values are: NONE OWNERS_ONLY MANAGERS_ONLY OWNERS_AND_MANAGERS ALL_MEMBERS

allow_external_members: Option<String>

Are external members allowed to join the group.

who_can_enter_free_form_tags: Option<String>

Permission to enter free form tags for topics in a forum. Possible values are: NONE OWNERS_ONLY MANAGERS_ONLY OWNERS_AND_MANAGERS ALL_MEMBERS

who_can_approve_messages: Option<String>

Permission to approve pending messages in the moderation queue. Possible values are: NONE OWNERS_ONLY OWNERS_AND_MANAGERS ALL_MEMBERS

who_can_mark_duplicate: Option<String>

Permission to mark a topic as a duplicate of another topic. Possible values are: NONE OWNERS_ONLY MANAGERS_ONLY OWNERS_AND_MANAGERS ALL_MEMBERS

who_can_join: Option<String>

Permissions to join the group. Possible values are: ANYONE_CAN_JOIN ALL_IN_DOMAIN_CAN_JOIN INVITED_CAN_JOIN CAN_REQUEST_TO_JOIN

who_can_modify_tags_and_categories: Option<String>

Permission to change tags and categories. Possible values are: NONE OWNERS_ONLY MANAGERS_ONLY OWNERS_AND_MANAGERS ALL_MEMBERS

who_can_mark_no_response_needed: Option<String>

Permission to mark a topic as not needing a response. Possible values are: NONE OWNERS_ONLY MANAGERS_ONLY OWNERS_AND_MANAGERS ALL_MEMBERS

who_can_unmark_favorite_reply_on_any_topic: Option<String>

Permission to unmark any post from a favorite reply. Possible values are: NONE OWNERS_ONLY MANAGERS_ONLY OWNERS_AND_MANAGERS ALL_MEMBERS

who_can_moderate_content: Option<String>

Permission for content moderation. Possible values are: NONE OWNERS_ONLY OWNERS_AND_MANAGERS ALL_MEMBERS

primary_language: Option<String>

Primary language for the group.

who_can_mark_favorite_reply_on_own_topic: Option<String>

Permission to mark a post for a topic they started as a favorite reply. Possible values are: NONE OWNERS_ONLY MANAGERS_ONLY OWNERS_AND_MANAGERS ALL_MEMBERS

who_can_view_membership: Option<String>

Permissions to view membership. Possible values are: ALL_IN_DOMAIN_CAN_VIEW ALL_MEMBERS_CAN_VIEW ALL_MANAGERS_CAN_VIEW ALL_OWNERS_CAN_VIEW

favorite_replies_on_top: Option<String>

If favorite replies should be displayed above other replies.

who_can_mark_favorite_reply_on_any_topic: Option<String>

Permission to mark any other user's post as a favorite reply. Possible values are: NONE OWNERS_ONLY MANAGERS_ONLY OWNERS_AND_MANAGERS ALL_MEMBERS

include_custom_footer: Option<String>

Whether to include custom footer.

who_can_move_topics_out: Option<String>

Permission to move topics out of the group or forum. Possible values are: NONE OWNERS_ONLY OWNERS_AND_MANAGERS ALL_MEMBERS

default_message_deny_notification_text: Option<String>

Default message deny notification message

include_in_global_address_list: Option<String>

If this groups should be included in global address list or not.

archive_only: Option<String>

If the group is archive only

who_can_delete_topics: Option<String>

Permission to delete topics. Possible values are: NONE OWNERS_ONLY OWNERS_AND_MANAGERS ALL_MEMBERS

who_can_delete_any_post: Option<String>

Permission to delete replies to topics. Possible values are: NONE OWNERS_ONLY OWNERS_AND_MANAGERS ALL_MEMBERS

is_archived: Option<String>

If the contents of the group are archived.

members_can_post_as_the_group: Option<String>

Can members post using the group email address.

who_can_make_topics_sticky: Option<String>

Permission to make topics appear at the top of the topic list. Possible values are: NONE OWNERS_ONLY MANAGERS_ONLY OWNERS_AND_MANAGERS ALL_MEMBERS

custom_roles_enabled_for_settings_to_be_merged: Option<String>

If any of the settings that will be merged have custom roles which is anything other than owners, managers, or group scopes.

email: Option<String>

Email id of the group

who_can_discover_group: Option<String>

Permission for who can discover the group. Possible values are: ALL_MEMBERS_CAN_DISCOVER ALL_IN_DOMAIN_CAN_DISCOVER ANYONE_CAN_DISCOVER

who_can_modify_members: Option<String>

Permission to modify members (change member roles). Possible values are: NONE OWNERS_ONLY OWNERS_AND_MANAGERS ALL_MEMBERS

message_moderation_level: Option<String>

Moderation level for messages. Possible values are: MODERATE_ALL_MESSAGES MODERATE_NON_MEMBERS MODERATE_NEW_MEMBERS MODERATE_NONE

description: Option<String>

Description of the group

who_can_unassign_topic: Option<String>

Permission to unassign any topic in a forum. Possible values are: NONE OWNERS_ONLY MANAGERS_ONLY OWNERS_AND_MANAGERS ALL_MEMBERS

reply_to: Option<String>

Whome should the default reply to a message go to. Possible values are: REPLY_TO_CUSTOM REPLY_TO_SENDER REPLY_TO_LIST REPLY_TO_OWNER REPLY_TO_IGNORE REPLY_TO_MANAGERS

custom_reply_to: Option<String>

Default email to which reply to any message should go.

send_message_deny_notification: Option<String>

Should the member be notified if his message is denied by owner.

enable_collaborative_inbox: Option<String>

If a primary Collab Inbox feature is enabled.

who_can_contact_owner: Option<String>

Permission to contact owner of the group via web UI. Possible values are: ANYONE_CAN_CONTACT ALL_IN_DOMAIN_CAN_CONTACT ALL_MEMBERS_CAN_CONTACT ALL_MANAGERS_CAN_CONTACT

message_display_font: Option<String>

Default message display font. Possible values are: DEFAULT_FONT FIXED_WIDTH_FONT

who_can_leave_group: Option<String>

Permission to leave the group. Possible values are: ALL_MANAGERS_CAN_LEAVE ALL_OWNERS_CAN_LEAVE ALL_MEMBERS_CAN_LEAVE NONE_CAN_LEAVE

who_can_add: Option<String>

Permissions to add members. Possible values are: ALL_MANAGERS_CAN_ADD ALL_OWNERS_CAN_ADD ALL_MEMBERS_CAN_ADD NONE_CAN_ADD

who_can_post_message: Option<String>

Permissions to post messages to the group. Possible values are: NONE_CAN_POST ALL_MANAGERS_CAN_POST ALL_MEMBERS_CAN_POST ALL_OWNERS_CAN_POST ALL_IN_DOMAIN_CAN_POST ANYONE_CAN_POST

who_can_move_topics_in: Option<String>

Permission to move topics into the group or forum. Possible values are: NONE OWNERS_ONLY OWNERS_AND_MANAGERS ALL_MEMBERS

who_can_take_topics: Option<String>

Permission to take topics in a forum. Possible values are: NONE OWNERS_ONLY MANAGERS_ONLY OWNERS_AND_MANAGERS ALL_MEMBERS

name: Option<String>

Name of the Group

kind: Option<String>

The type of the resource.

max_message_bytes: Option<i32>

Maximum message size allowed.

who_can_invite: Option<String>

Permissions to invite members. Possible values are: ALL_MEMBERS_CAN_INVITE ALL_MANAGERS_CAN_INVITE ALL_OWNERS_CAN_INVITE NONE_CAN_INVITE

who_can_approve_members: Option<String>

Permission to approve members. Possible values are: ALL_OWNERS_CAN_APPROVE ALL_MANAGERS_CAN_APPROVE ALL_MEMBERS_CAN_APPROVE NONE_CAN_APPROVE

spam_moderation_level: Option<String>

Moderation level for messages detected as spam. Possible values are: ALLOW MODERATE SILENTLY_MODERATE REJECT

allow_web_posting: Option<String>

If posting from web is allowed.

who_can_moderate_members: Option<String>

Permission for membership moderation. Possible values are: NONE OWNERS_ONLY OWNERS_AND_MANAGERS ALL_MEMBERS

who_can_add_references: Option<String>

Permission to add references to a topic. Possible values are: NONE OWNERS_ONLY MANAGERS_ONLY OWNERS_AND_MANAGERS ALL_MEMBERS

who_can_view_group: Option<String>

Permissions to view group. Possible values are: ANYONE_CAN_VIEW ALL_IN_DOMAIN_CAN_VIEW ALL_MEMBERS_CAN_VIEW ALL_MANAGERS_CAN_VIEW ALL_OWNERS_CAN_VIEW

show_in_group_directory: Option<String>

Is the group listed in groups directory

who_can_post_announcements: Option<String>

Permission to post announcements, a special topic type. Possible values are: NONE OWNERS_ONLY OWNERS_AND_MANAGERS ALL_MEMBERS

who_can_lock_topics: Option<String>

Permission to lock topics. Possible values are: NONE OWNERS_ONLY OWNERS_AND_MANAGERS ALL_MEMBERS

who_can_assign_topics: Option<String>

Permission to assign topics in a forum to another user. Possible values are: NONE OWNERS_ONLY MANAGERS_ONLY OWNERS_AND_MANAGERS ALL_MEMBERS

custom_footer_text: Option<String>

Custom footer text.

allow_google_communication: Option<String>

Is google allowed to contact admins.

who_can_hide_abuse: Option<String>

Permission to hide posts by reporting them as abuse. Possible values are: NONE OWNERS_ONLY MANAGERS_ONLY OWNERS_AND_MANAGERS ALL_MEMBERS