Struct google_groupssettings1::Groups

pub struct Groups {
    pub message_moderation_level: Option<String>,
    pub description: Option<String>,
    pub allow_external_members: Option<String>,
    pub reply_to: Option<String>,
    pub custom_reply_to: Option<String>,
    pub send_message_deny_notification: Option<String>,
    pub who_can_contact_owner: Option<String>,
    pub message_display_font: Option<String>,
    pub is_archived: Option<String>,
    pub who_can_post_message: Option<String>,
    pub primary_language: Option<String>,
    pub who_can_view_membership: Option<String>,
    pub who_can_invite: Option<String>,
    pub who_can_join: Option<String>,
    pub kind: Option<String>,
    pub name: Option<String>,
    pub spam_moderation_level: Option<String>,
    pub default_message_deny_notification_text: Option<String>,
    pub who_can_view_group: Option<String>,
    pub include_in_global_address_list: Option<String>,
    pub archive_only: Option<String>,
    pub show_in_group_directory: Option<String>,
    pub who_can_leave_group: Option<String>,
    pub members_can_post_as_the_group: Option<String>,
    pub max_message_bytes: Option<i32>,
    pub allow_web_posting: Option<String>,
    pub email: Option<String>,
    pub allow_google_communication: Option<String>,
}

JSON template for Group resource

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

update groups (request|response)

patch groups (request|response)

get groups (response)

Fields

message_moderation_level

Moderation level for messages. Possible values are: MODERATE_ALL_MESSAGES MODERATE_NON_MEMBERS MODERATE_NEW_MEMBERS MODERATE_NONE

description

Description of the group

allow_external_members

Are external members allowed to join the group.

reply_to

Whome should the default reply to a message go to. Possible values are: REPLY_TO_CUSTOM REPLY_TO_SENDER REPLY_TO_LIST REPLY_TO_OWNER REPLY_TO_IGNORE REPLY_TO_MANAGERS

custom_reply_to

Default email to which reply to any message should go.

send_message_deny_notification

Should the member be notified if his message is denied by owner.

who_can_contact_owner

Permission to contact owner of the group via web UI. Possbile values are: ANYONE_CAN_CONTACT ALL_IN_DOMAIN_CAN_CONTACT ALL_MEMBERS_CAN_CONTACT ALL_MANAGERS_CAN_CONTACT

message_display_font

Default message display font. Possible values are: DEFAULT_FONT FIXED_WIDTH_FONT

is_archived

If the contents of the group are archived.

who_can_post_message

Permissions to post messages to the group. Possible values are: NONE_CAN_POST ALL_MANAGERS_CAN_POST ALL_MEMBERS_CAN_POST ALL_IN_DOMAIN_CAN_POST ANYONE_CAN_POST

primary_language

Primary language for the group.

who_can_view_membership

Permissions to view membership. Possbile values are: ALL_IN_DOMAIN_CAN_VIEW ALL_MEMBERS_CAN_VIEW ALL_MANAGERS_CAN_VIEW

who_can_invite

Permissions to invite members. Possbile values are: ALL_MEMBERS_CAN_INVITE ALL_MANAGERS_CAN_INVITE

who_can_join

Permissions to join the group. Possible values are: ANYONE_CAN_JOIN ALL_IN_DOMAIN_CAN_JOIN INVITED_CAN_JOIN CAN_REQUEST_TO_JOIN

kind

The type of the resource.

name

Name of the Group

spam_moderation_level

Moderation level for messages detected as spam. Possible values are: ALLOW MODERATE SILENTLY_MODERATE REJECT

default_message_deny_notification_text

Default message deny notification message

who_can_view_group

Permissions to view group. Possbile values are: ANYONE_CAN_VIEW ALL_IN_DOMAIN_CAN_VIEW ALL_MEMBERS_CAN_VIEW ALL_MANAGERS_CAN_VIEW

include_in_global_address_list

If this groups should be included in global address list or not.

archive_only

If the group is archive only

show_in_group_directory

Is the group listed in groups directory

who_can_leave_group

Permission to leave the group. Possible values are: ALL_MANAGERS_CAN_LEAVE ALL_MEMBERS_CAN_LEAVE

members_can_post_as_the_group

Can members post using the group email address.

max_message_bytes

Maximum message size allowed.

allow_web_posting

If posting from web is allowed.

email

Email id of the group

allow_google_communication

Is google allowed to contact admins.