Struct google_groupssettings1::Groups

pub struct Groups {
    pub send_message_deny_notification: Option<String>,
    pub description: Option<String>,
    pub allow_external_members: Option<String>,
    pub reply_to: Option<String>,
    pub custom_reply_to: Option<String>,
    pub message_moderation_level: Option<String>,
    pub who_can_contact_owner: Option<String>,
    pub message_display_font: Option<String>,
    pub is_archived: Option<String>,
    pub who_can_add: Option<String>,
    pub who_can_post_message: Option<String>,
    pub primary_language: Option<String>,
    pub who_can_view_membership: Option<String>,
    pub who_can_invite: Option<String>,
    pub who_can_join: Option<String>,
    pub kind: Option<String>,
    pub name: Option<String>,
    pub include_custom_footer: Option<String>,
    pub spam_moderation_level: Option<String>,
    pub custom_footer_text: Option<String>,
    pub default_message_deny_notification_text: Option<String>,
    pub who_can_view_group: Option<String>,
    pub include_in_global_address_list: Option<String>,
    pub archive_only: Option<String>,
    pub show_in_group_directory: Option<String>,
    pub who_can_leave_group: Option<String>,
    pub members_can_post_as_the_group: Option<String>,
    pub max_message_bytes: Option<i32>,
    pub allow_web_posting: Option<String>,
    pub email: Option<String>,
    pub allow_google_communication: Option<String>,
}

JSON template for Group resource

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

update groups (request|response)

patch groups (request|response)

get groups (response)

Fields
send_message_deny_notification: Option<String>

Should the member be notified if his message is denied by owner.

description: Option<String>

Description of the group

allow_external_members: Option<String>

Are external members allowed to join the group.

reply_to: Option<String>

Whome should the default reply to a message go to. Possible values are: REPLY_TO_CUSTOM REPLY_TO_SENDER REPLY_TO_LIST REPLY_TO_OWNER REPLY_TO_IGNORE REPLY_TO_MANAGERS

custom_reply_to: Option<String>

Default email to which reply to any message should go.

message_moderation_level: Option<String>

Moderation level for messages. Possible values are: MODERATE_ALL_MESSAGES MODERATE_NON_MEMBERS MODERATE_NEW_MEMBERS MODERATE_NONE

who_can_contact_owner: Option<String>

Permission to contact owner of the group via web UI. Possible values are: ANYONE_CAN_CONTACT ALL_IN_DOMAIN_CAN_CONTACT ALL_MEMBERS_CAN_CONTACT ALL_MANAGERS_CAN_CONTACT

message_display_font: Option<String>

Default message display font. Possible values are: DEFAULT_FONT FIXED_WIDTH_FONT

is_archived: Option<String>

If the contents of the group are archived.

who_can_add: Option<String>

Permissions to add members. Possible values are: ALL_MANAGERS_CAN_ADD ALL_MEMBERS_CAN_ADD NONE_CAN_ADD

who_can_post_message: Option<String>

Permissions to post messages to the group. Possible values are: NONE_CAN_POST ALL_MANAGERS_CAN_POST ALL_MEMBERS_CAN_POST ALL_IN_DOMAIN_CAN_POST ANYONE_CAN_POST

primary_language: Option<String>

Primary language for the group.

who_can_view_membership: Option<String>

Permissions to view membership. Possible values are: ALL_IN_DOMAIN_CAN_VIEW ALL_MEMBERS_CAN_VIEW ALL_MANAGERS_CAN_VIEW

who_can_invite: Option<String>

Permissions to invite members. Possible values are: ALL_MEMBERS_CAN_INVITE ALL_MANAGERS_CAN_INVITE NONE_CAN_INVITE

who_can_join: Option<String>

Permissions to join the group. Possible values are: ANYONE_CAN_JOIN ALL_IN_DOMAIN_CAN_JOIN INVITED_CAN_JOIN CAN_REQUEST_TO_JOIN

kind: Option<String>

The type of the resource.

name: Option<String>

Name of the Group

include_custom_footer: Option<String>

Whether to include custom footer.

spam_moderation_level: Option<String>

Moderation level for messages detected as spam. Possible values are: ALLOW MODERATE SILENTLY_MODERATE REJECT

custom_footer_text: Option<String>

Custom footer text.

default_message_deny_notification_text: Option<String>

Default message deny notification message

who_can_view_group: Option<String>

Permissions to view group. Possible values are: ANYONE_CAN_VIEW ALL_IN_DOMAIN_CAN_VIEW ALL_MEMBERS_CAN_VIEW ALL_MANAGERS_CAN_VIEW

include_in_global_address_list: Option<String>

If this groups should be included in global address list or not.

archive_only: Option<String>

If the group is archive only

show_in_group_directory: Option<String>

Is the group listed in groups directory

who_can_leave_group: Option<String>

Permission to leave the group. Possible values are: ALL_MANAGERS_CAN_LEAVE ALL_MEMBERS_CAN_LEAVE NONE_CAN_LEAVE

members_can_post_as_the_group: Option<String>

Can members post using the group email address.

max_message_bytes: Option<i32>

Maximum message size allowed.

allow_web_posting: Option<String>

If posting from web is allowed.

email: Option<String>

Email id of the group

allow_google_communication: Option<String>

Is google allowed to contact admins.